+++ src/main/java/net/frozenblock/trailiertales/mixin/common/apparition/MobMixin.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2025 FrozenBlock
+ * This file is part of Trailier Tales.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.frozenblock.trailiertales.mixin.common.apparition;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import net.frozenblock.trailiertales.config.TTEntityConfig;
+import net.frozenblock.trailiertales.entity.Apparition;
+import net.minecraft.world.entity.Mob;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(Mob.class)
+public class MobMixin {
+
+	@ModifyExpressionValue(
+		method = "aiStep",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/GameRules;getBoolean(Lnet/minecraft/world/level/GameRules$Key;)Z"
+		)
+	)
+	private boolean bruh(boolean original) {
+		if (original) return true;
+		Mob mob	= (Mob) (Object) this;
+		if (mob instanceof Apparition) {
+			return TTEntityConfig.get().apparition.ignore_mob_griefing;
+		}
+		return false;
+	}
+}