+++ CHANGELOG.md
@@ -3,5 +3,4 @@ Thank you!
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 hi
 -----------------
-- Fixed players being able to place Ectoplasm Blocks when they shouldn't be able to.
-- Fixed the placement of the Ectoplasm Block and Polished Resin blocks in the Creative inventory.
+- Added a config option to allow Apparitions to ignore the mobGriefing gamerule